added code to wait if disk ejection is disabled
Added a line of code to continuously wait 5 seconds, then poll the drive and repeat if the disk is not open or empty. This will prevent ripper from unintentionally ripping the same disk over and over again if auto eject is disabled. there is a potential concern that if the user removes the old disk and inserts a new one within 5 seconds that it doesn't get recognised. That can be resolved by waiting longer before inserting a disk. It may be possible to also add a check against if a disk is "loading",  but I don't know if an idle disk would randomly load which will cancel the wait unintentionally.